Move sketch solve machine coverage to integration
